finish register
