+++ bingus/src/app/page.tsx
@@ -1,16 +1,9 @@
 "use server";
-import RegisterPage from "./register/RegisterPage";
 
 export default async function Page() {
-
-
-
   return (
     <div>
-      <RegisterPage />      
-      
+      Home Page
     </div>
-
-    
   );
 }
