+++ bingus/src/app/register/error.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+export default function Error({
+    error,
+    reset,
+}: {
+    error: Error & { digest?: string }
+    reset: () => void
+}) {
+    return (
+        <div>
+            Error occured.
+            {error.stack}
+        </div>
+    );
+}
