+++ bingus/src/public/logo.jpg
